Preparation 2

+++ Select Lucky Dog.Core/Models/Student.cs
@@ -8,10 +8,10 @@ namespace Select_Lucky_Dog.Core.Models
     {
         public string Name { get; set; }
         public StudentStatus Status { get; set; }
-        public byte OrderOfGoing { get; set; }
-        public byte OrderInList { get; set; }
+        public int OrderOfGoing { get; set; }
+        public int OrderInList { get; set; }
         public override string ToString() => Name + "\t" + Status.ToString() + "\t" + (Status == StudentStatus.going ? OrderOfGoing.ToString() : "");
-        public Student(string name,StudentStatus status,byte orderOfGoing,byte orderInList)
+        public Student(string name,StudentStatus status,int orderOfGoing,int orderInList)
         {
             this.Name=name;
             this.Status=status;
